move indexing behavior into a mixin
